Backend + browser code/requirements.txt
flask
flask-cors
requests
beautifulsoup4
tqdm
openai
lxml
python-dotenv
bs4
gunicorn
itsdangerous
